requirements.txt
# Healthcare Eligibility Pipeline - Dependencies
# Python 3.9+

# Core data processing
pandas==2.1.4

# Date parsing (used in helpers.py)
python-dateutil==2.8.2

# Configuration management
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0